Update to use extension and along with other changes needed
