+++ src/AssemblySharedInfoGenerator/AssemblySharedInfo.tt
@@ -9,7 +9,7 @@ using System.Runtime.InteropServices;
 // associated with an assembly.
 [assembly: AssemblyCompany("Autodesk, Inc")]
 [assembly: AssemblyProduct("Dynamo")]
-[assembly: AssemblyCopyright("Copyright © Autodesk, Inc 2017")]
+[assembly: AssemblyCopyright("Copyright © Autodesk, Inc 2020")]
 [assembly: AssemblyTrademark("")]
 
 //In order to begin building localizable applications, set
