+++ src/IronPythonExtension/Properties/AssemblyInfo.cs
@@ -1,0 +1,11 @@
+﻿using System.Reflection;
+using System.Runtime.InteropServices;
+
+// General Information about an assembly is controlled through the following
+// set of attributes. Change these attribute values to modify the information
+// associated with an assembly.
+[assembly: AssemblyTitle("IronPythonExtension")]
+[assembly: AssemblyCulture("")]
+
+// The following GUID is for the ID of the typelib if this project is exposed to COM
+[assembly: Guid("182fca4e-b6ef-451f-9ec4-7bf2c622f4f7")]
