+++ src/PythonExtension/IronPythonExtension_ExtensionDefinition.xml
@@ -1,0 +1,4 @@
+<ViewExtensionDefinition>
+  <AssemblyPath>..\IronPythonExtension.dll</AssemblyPath>
+  <TypeName>IronPythonExtension.IronPythonExtension</TypeName>
+</ViewExtensionDefinition>
