--- src/PythonExtension/PythonExtension_ViewExtensionDefinition.xml
@@ -1,4 +1,0 @@
-<ViewExtensionDefinition>
-  <AssemblyPath>..\PythonExtension.dll</AssemblyPath>
-  <TypeName>PythonExtension.PythonViewExtension</TypeName>
-</ViewExtensionDefinition>
